updated list layout added filter form and logic, updated error and success messaging in addItem
