--- src/components/Errors.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Errors({ message }) {
-  return <div className="error">Error: {message}</div>;
-}
-
-export default Errors;
